requirements.txt
numpy
neo4j
openai
python-dotenv
tqdm
lxml
Flask
Flask-Login
Flask-WTF
WTForms
Werkzeug
Flask-SQLAlchemy
